backend/requirements-api.txt
# Dépendances Backend FastAPI
fastapi==0.129.0
uvicorn==0.41.0
starlette==0.52.1

pydantic==2.12.5
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2

python-multipart==0.0.22
requests==2.32.4
httpx==0.28.1
anyio==4.11.0
h11==0.16.0
idna==3.10
sniffio==1.3.1

PyYAML==6.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

cryptography==46.0.5
cffi==2.0.0
pycparser==2.23
google-genai==1.64.0

# PDF / OCR 
PyMuPDF==1.26.0
pytesseract==0.3.13

# utils
tqdm==4.67.1